backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.6

# File Parsing
python-docx==1.1.2
openpyxl==3.1.5
PyMuPDF==1.25.3

# Google AI
google-genai==1.14.0

# Export
xlsxwriter==3.2.0

# Utilities
python-dotenv==1.0.1
pydantic[email]==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
aiofiles==24.1.0
